requirements.txt
attrs==23.1.0
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==42.0.5
cssselect==1.2.0
filelock==3.12.4
greenlet==3.0.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
packaging==23.1
parsel==1.8.1
playwright==1.42.0
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.1
scrapy-playwright==0.0.32
scrapy-splash==0.9.0
service-identity==23.1.0
six==1.16.0
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.8.0
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.0
opencv-python==4.9.0.80
scikit-image==0.22.0
# Flask dependencies
Django==5.0.3
djangorestframework==3.15.0
django-filter==24.1